requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
timm==0.9.12
pillow==10.1.0
numpy==1.26.2
